Changed Point System
I adjusted the point system from an improvised and pseudo-accurate math equation to a more precise approximation on the part of ratio of fee to amount inputted.

I also added a feature that allows you to see the minimum growth by amount that you'd need to wait till you break even including the fees. This'll help in the event you want a safe spot and don't know how much to take out.

Am currently working on a rounding equation that'll allow me to round up from a certain decimal place in order to ensure that I can always pay/expect the right fee. Otherwise, if we round down, there'll be losses to how much we're supposed to put in.

But yeah other than that there hasn't been too many changes. Still the same goal. Planning to implement a history feature that allows you to track current orders